[refactoring] Remove code duplication for nvcf client classes (#1334)
Define base class `BaseNvcfRuntimeClient` from which `NvcfRuntimeClient` (C++) and `PyNvcfRuntimeClient` (Python) inherit
so as to remove the duplicate code across files / classes.
